Move the Rails test helper into lib/scaffolding_extensions directory
This makes it so you can just:

  require 'scaffolding_extensions/rails_test_help'

To get the test helper methods.  Additionally, it adds a
scaffold_session method that can be modified to specify a session
hash to use for the page requests.  This is necessary if the scaffold
pages require a login, which they generally do, considering that
scaffolding_extensions is mainly useful for administrative purposes.

Bump version to 1.1.8.