Update test site to test with Rails/ActiveRecord 2.2.2
Rails 2.2.2 causes an issue with DataMapper that none of the other
frameworks seem to experience, which is duplicate entries in
scaffold_fields.  I'm guessing it's caused somewhere by
ActiveSupport or it's a threading issue. I can't replicate it in the
console.  Work around the issue by using uniq in the DataMapper
adapter's scaffold_fields.

ActiveRecord 2.2.2 is very slow when used on the other
frameworks, but not on Rails.  On the other frameworks,
ActiveRecord is just required and used, while Rails sets it
up when using Rails, so I'm guessing that ActiveRecord 2.2.2
needs some extra configuration that I'm not providing in order
to go fast.

Update README. No version bumping because it's unlikely there are
any users using scaffolding extensions with Rails 2.2.2 and
DataMapper 0.3.