Remove support for DataMapper
ScaffoldingExtensions only supported the old version of DataMapper,
which is unlikely to be used by anyone.  Due to the "great
refactoring" of DataMapper, it's unlikely the current support would
help in implementing support for the new version.  If someone is
interested, it shouldn't be too difficult to add support for the
current version of DataMapper, but I have no interest in doing so
myself.

This was specifically done after the addition of Merb support,
so anyone using Merb and the old version of DataMapper can still
use git to get a version of scaffolding extensions that supports
both.

While here, remove setup_test_databases, as bootstrap.rb does
the same thing for the test site.