Full support on Sinatra, scaffold_session_id bugfix, bump version to 1.3.1
This commit mainly improves support on Sinatra.  Before, Sinatra
support was really a hack, by setting up our own anonymous controller
classes and adding the scaffolding methods to those.  That worked but
didn't integrate well with Sinatra.

Now, we use the Sinatra::EventContext class as the controller class,
which integrates well and allows us to override the scaffold
templates and layout with versions provided by the user.  One
downside to this is that Sinatra::EventContext is a single class,
which means you can't scaffold different model classes with the same
scaffold_name in a single sinatra instance, even if you mount the
scaffolds on a different path.  You can work around this by making
sure all models have a unique scaffold_name.  This is unlikely to
affect anything but the test_site.

This commit fixes a bug in scaffold_unassociated_objects that used
the wrong scaffold_session_value parameter.  This affected the habtm
scaffolds when the current class and the associated class did not
have the same scaffold_session_value.

Fix a bug in the the Sinatra support by checking that the method
requested is a scaffolded method before sending it.  This may be
at least a DOS level bug (i.e. fork).  I don't think it has security
implications, as no arguments were sent with the method, but if you
defined an Object or Kernel method that took no arguments, and
had security implications, you may have been vulnerable.

Update merb and camping docs to reflect the level of support.

Fix test site to test sinatra separately for sequel and
active_record.

Bump version to 1.3.1.