Don't use wierd camping style redirect in Sinatra adapter, bump version to 1.3.3
This makes the Sinatra adapter use a simple path instead of the
weird //server/path redirect that camping does.  This appears to
work, though it does violate the RFC (the camping style redirect
violates the RFC as well).  You can use the Rack::RelativeRedirect
middleware in rack-contrib to convert relative paths in redirects
to absolute URLs.