Allow :only and :except options for scaffold_all_models to be underscored strings, bump version to 1.3.5
This changes ScaffoldingExtensions.all_models to take options.  The
main usage will be the ability to add underscored model name strings
to :except, to avoid referencing a constant that doesn't exist. For
example, if you have a file named init.rb in your models directory,
all_models will try to reference a constant named Init, which will
give an error if it doesn't exist.  The only workaround for this
previously was to specify all models to use via :only, which was
tedious.

This also removes scaffold_normalize_options, which basically
duplicated Kernel#Array.