Small bug fixes and improvements on sinatra
The main fix is having the scaffolding extensions route match "/".
Before it would match "", but not "/".

The second fix is that scaffolding extensions will now pass
instead of raising Sinatra::NotFound if the route matches but
there isn't a method defined.  I still encourage making your
scaffold calls the last method calls if you are using Sinatra.

The third fix is that scaffolding extensions will now raise an
error if a HEAD request is issued for a nonidempotent method.

This also uses REQUEST_METHOD from the env instead of using a
fairly roundabout way to determine it.

Thie also only defines the route block once to save on memory.
It adds a separate get route for "" and "/".