In list scaffold, assume only show and edit actions are idempotent
Previously, it assumed that all actions but destroy were idempotent.
Now, scaffolding extensions only supports show, edit, and destroy by
default, but it's fairly easy to extend to support other actions,
and it's safer to assume that those methods won't be idempotent.

This adds the scaffold_idempotent_method? helper method which can
be overridden if you want to add custom idempotent methods that
use get instead of post.