Support Ramaze 2009.06.04
Make scaffolding_extensions fully support the Innate-based versions
of Ramaze.  Unfortunately, full support requires a change just
made to the git version of Innate, but it should be mostly working
even on the gem version.

Change the code so that all frameworks use [] as a suffix for
parameters that should be a list.  This apparently is backwards
compatible, and is required when using rack 1.0.0 (which ramaze
indirectly depends on).

Refactor scaffold_add_associated_objects and makes sure that
associated objects don't get added multiple times when using
Sequel.

Remove scaffold_force_array, since it just duplicates the
Kernel#Array method poorly.

Move the CleanUpARGarbage middleware into a separate file that
can be used by both Ramaze and Sinatra, since Ramaze's after_all
block doesn't seem to work as before.

While here, bump supported Sequel version to 3.1.0.

The gem version isn't getting bumped yet, it will after a gem
version of Innate with the fix is released.