Use instance_variable_defined? in Overridable so SE works on 1.8 and 1.9
I was on the fence about using instance_variable_defined?, since it
was added in 1.8.6, and there are a lot of systems still using 1.8.5
(CentOS/RHEL 5).  However, I bumped the required ruby version of the
gem to 1.8.6, so people using 1.8.5 shouldn't be able to install it.
If there is anyone using 1.8.5 that wants to use future SE versions, I
don't have a problem modifying things.