Fix pagination of search results page when searching by field, bump version to 1.3.10
This has probably been broken for a long time (my guess is since
February 2008), but was only reported to me today.  The problem is
the hidden form fields on the search resutls page use the old
instance variables instead of the new scaffold_options hash.  Fix
that and add tests for it.

While here, correctly close tags for next and previous page buttons.