Support Rails 2.3.5 with or without the RailsXss plugin, bump version to 1.3.12
Rails 2.3.5 was already compatible without the RailsXss plugin, but
this commit allows scaffolding_extensions to work with the RailsXss
and make sure that HTML is not doubly escaped.

It was fairly simple to handle most of the work, using the
safe_helper metaprogramming method that RailsXss gives you.  There
were some tedious changes to the scaffold templates so that they
would work both with and without RailsXss.