Add a render_* controller level method for allowing custom controller actions while still using SE for rendering
Previously, it was quite difficult to override the scaffolded
methods and still use scaffolding_extensions for rendering.
This is because the rendering was done by
scaffold_render_template, which took a hash of options that
the user didn't have direct access to, since it was defined
as a local variable in a class method and just used in the
instance method closures.

This commit adds a render_* private controller method (e.g.
render_album) that just calls the scaffold_render_template
method with the correct hash so that the scaffolded forms
work correctly.

Here's an example of how you could use it:

  # Override
  def list_album
    # Only show albums that haven't been deleted, instead of
    # all albums.
    @scaffold_objects = Album.filter(:deleted=>false).all
    render_album(:list)
  end