Add support for Rails 3.0.0, drop support for Rails 2.3
Most of the changes are adding scaffold_raw to all of the scaffolds
instead of using safe_helper on the internal methods.  This will
make it easier to support other web frameworks if they want to use
similar XSS support.

The scaffold layout support in Rails 3 is uglier than in previous
versions as render_to_string in Rails 3 doesn't add the instance
variables back to the controller like it did in Rails 2.3.