Remove Merb support
We only supported an old version of Merb (1.0.4), and it was only ever
tested against a patched version.  I can't recall ever getting asked
about it, and since Merb users are encouraged to upgrade to Rails 3,
which scaffolding_extensions already supports, there doesn't seem to
be a reason to keep this.  The reason I am dropping support now is
that the patched version of 1.0.4 I was using doesn't support 1.9.2.